# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ChangePasswordRequest": {
        "description": "Request model for changing password.",
        "properties": {
          "currentPassword": {
            "title": "Currentpassword",
            "type": "string"
          },
          "newPassword": {
            "title": "Newpassword",
            "type": "string"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ],
        "title": "ChangePasswordRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UpdateProfileRequest": {
        "description": "Request model for updating user profile.",
        "properties": {
          "displayName": {
            "title": "Displayname",
            "type": "string"
          }
        },
        "required": [
          "displayName"
        ],
        "title": "UpdateProfileRequest",
        "type": "object"
      },
...
      "UserCreateRequest": {
        "properties": {
          "displayName": {
            "title": "Displayname",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "viewer",
            "title": "Role"
          }
        },
        "required": [
          "displayName",
          "email",
          "password"
        ],
        "title": "UserCreateRequest",
        "type": "object"
      },
...
      "UserResponse": {
        "properties": {
          "displayName": {
            "title": "Displayname",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "displayName",
          "role"
        ],
        "title": "UserResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "REST API for Open Projects Hub \u2014 project, client, and story management with JWT authentication and role-based access control. See [docs/api/README.md](https://github.com/alonsovndev/open-projects-hub-api/blob/main/docs/api/README.md) for auth flows, pagination, and code examples.",
    "title": "Open Projects Hub API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/users": {
      "post": {
        "description": "Create a new user (admin only).\n\nRequires ADMIN role. For public self-registration, use POST /v1/auth/register instead.\nAdmins can specify the role (admin or viewer) when creating users.\nDefaults to viewer if not specified.\n\nArgs:\n    payload: UserCreateRequest with email, password, displayName, optional role\n    create_user_use_case: Injected CreateUserUseCase\n    current_user: Current authenticated admin user\n\nReturns:\n    UserResponse with created user data\n\nRaises:\n    403: Forbidden (non-admin user)\n    409: Conflict (email already exists)\n    422: Validation error (invalid payload)",
        "operationId": "create_user_v1_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create User",
        "tags": [
          "Users"
        ]
      }
    },
    "/v1/users/me/password": {
      "post": {
        "description": "Change user password.\n\nRequires authentication. Verifies the current password before updating\nto the new password. New password must meet complexity requirements.\n\nArgs:\n    payload: ChangePasswordRequest with currentPassword and newPassword\n    current_user: Current authenticated user (from JWT)\n    use_case: Injected ChangePasswordUseCase\n\nReturns:\n    Success message confirming password change\n\nRaises:\n    400: Validation failed (weak password, incorrect current password)\n    401: Unauthorized (invalid or missing token)\n    404: User not found\n    500: Internal server error",
        "operationId": "change_password_v1_users_me_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Change Password V1 Users Me Password Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Change Password",
        "tags": [
          "Users"
        ]
      }
    },
    "/v1/users/me/profile": {
      "get": {
        "description": "Get current user profile.\n\nRequires authentication. Returns profile for the authenticated user from the JWT token.\n\nArgs:\n    current_user: Current authenticated user (from JWT)\n    use_case: Injected GetUserProfileUseCase\n\nReturns:\n    UserResponse with profile data (id, email, displayName, role)\n\nRaises:\n    401: Unauthorized (invalid or missing token)\n    404: User not found\n    500: Internal server error",
        "operationId": "get_user_profile_v1_users_me_profile_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get User Profile",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "description": "Update current user profile.\n\nRequires authentication. Only the display_name field can be updated.\nEmail and role cannot be changed via this endpoint.\n\nArgs:\n    payload: UpdateProfileRequest with displayName\n    current_user: Current authenticated user (from JWT)\n    use_case: Injected UpdateUserProfileUseCase\n\nReturns:\n    UserResponse with updated profile data\n\nRaises:\n    400: Validation failed (empty name, too long, etc.)\n    401: Unauthorized (invalid or missing token)\n    404: User not found\n    500: Internal server error",
        "operationId": "update_user_profile_v1_users_me_profile_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update User Profile",
        "tags": [
          "Users"
        ]
      }
    },
    "/v1/users/{user_id}": {
      "get": {
        "description": "Get user by ID.\n\nRequires authentication.\n\nArgs:\n    user_id: User UUID\n    get_user_use_case: Injected GetUserByIdUseCase\n    current_user: Current authenticated user\n\nReturns:\n    UserResponse with user details (id, email, displayName, role)\n\nRaises:\n    401: Unauthorized\n    404: User not found\n    500: Internal server error",
        "operationId": "get_user_by_id_v1_users__user_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get User By Id",
        "tags": [
          "Users"
        ]
      }
    }
  }
}